+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_top_level_members_in_tests/examples/example.dart
@@ -1,0 +1,28 @@
+final public = 1; // LINT
+final _private = 2;
+
+void main() {}
+
+void function() {} // LINT
+
+void _function() {}
+
+class Class {} // LINT
+
+class _Class {}
+
+mixin Mixin {} // LINT
+
+mixin _Mixin {}
+
+extension Extension on String {} // LINT
+
+extension _Extension on String {}
+
+enum Enum { first, second } // LINT
+
+enum _Enum { first, second }
+
+typedef Public = String; // LINT
+
+typedef _Private = String;