+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * feat: add static code diagnostics [`prefer-correct-edge-insets-constructor-rule`](https://dartcodemetrics.dev/docs/rules/flutter/prefer-correct-edge-insets-constructor).
+* chore: restrict `analyzer` version to `>=2.7.0 <4.3.0`.
 
 ## 4.16.0