+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * feat: add static code diagnostic [`avoid-use-expanded-as-spacer`](https://dartcodemetrics.dev/docs/rules/flutter/avoid-use-expanded-as-spacer).
 * chore: restrict `analyzer` version to `>=2.7.0 <4.3.0`.
+* feat: add static code diagnostic [`avoid-duplicate-exports`](https://dartcodemetrics.dev/docs/rules/common/avoid-duplicate-exports).
 
 ## 4.16.0